+++ tests/linen/linen_test.py
@@ -114,9 +114,8 @@ class NormalizationTest(absltest.TestCase):
     x = random.normal(key1, (2, 3, 4))
     model_cls = nn.LayerNorm(use_bias=False, use_scale=False, epsilon=e)
     y, _ = model_cls.init_with_output(key2, x)
-    assert x.shape == y.shape
-    input_type = type(x)
-    assert isinstance(y, input_type)
+    self.assertEqual(x.dtype, y.dtype)
+    self.assertEqual(x.shape, y.shape)
     y_one_liner = ((x - x.mean(axis=-1, keepdims=True)) *
                    jax.lax.rsqrt(x.var(axis=-1, keepdims=True) + e))
     np.testing.assert_allclose(y_one_liner, y, atol=1e-4)
@@ -129,8 +128,8 @@ class NormalizationTest(absltest.TestCase):
     model_cls = nn.GroupNorm(num_groups=2, use_bias=False, use_scale=False, epsilon=e)
 
     y, _ = model_cls.init_with_output(key2, x)
+    self.assertEqual(x.dtype, y.dtype)
     self.assertEqual(x.shape, y.shape)
-    self.assertIsInstance(y, type(x))
 
     x_gr = x.reshape([2, 5, 4, 4, 2, 16])
     y_test = ((x_gr - x_gr.mean(axis=[1, 2, 3, 5], keepdims=True)) *